+++ core/amber/src/main/scala/edu/uci/ics/texera/web/resource/dashboard/user/workflow/WorkflowAccessResource.scala
@@ -2,12 +2,7 @@ package edu.uci.ics.texera.web.resource.dashboard.user.workflow
 
 import edu.uci.ics.texera.web.SqlServer
 import edu.uci.ics.texera.web.auth.SessionUser
-import edu.uci.ics.texera.web.model.common.AccessEntry2
-import edu.uci.ics.texera.web.model.jooq.generated.Tables.{
-  USER,
-  WORKFLOW_OF_USER,
-  WORKFLOW_USER_ACCESS
-}
+import edu.uci.ics.texera.web.model.jooq.generated.Tables.{USER, WORKFLOW_USER_ACCESS}
 import edu.uci.ics.texera.web.model.jooq.generated.enums.WorkflowUserAccessPrivilege
 import edu.uci.ics.texera.web.model.jooq.generated.tables.daos.{
   UserDao,
@@ -23,7 +18,6 @@ import org.jooq.types.UInteger
 import javax.annotation.security.RolesAllowed
 import javax.ws.rs._
 import javax.ws.rs.core.MediaType
-import scala.collection.convert.ImplicitConversions.`collection AsScalaIterable`
 
 object WorkflowAccessResource {
   final private val context: DSLContext = SqlServer.createDSLContext
@@ -101,45 +95,61 @@ class WorkflowAccessResource() {
 
   /**
     * Returns information about all current shared access of the given workflow
+    *
     * @param wid workflow id
-    * @return a List of email/permission pair
+    * @return a List of email/name/permission
     */
   @GET
   @Path("/list/{wid}")
   def getAccessList(
       @PathParam("wid") wid: UInteger,
       @Auth sessionUser: SessionUser
-  ): List[AccessEntry2] = {
-    val user = sessionUser.getUser
-    if (
-      workflowOfUserDao.existsById(
-        context
-          .newRecord(WORKFLOW_OF_USER.UID, WORKFLOW_OF_USER.WID)
-          .values(user.getUid, wid)
+  ) = {
+    context
+      .select(
+        USER.EMAIL,
+        USER.NAME,
+        WORKFLOW_USER_ACCESS.PRIVILEGE
       )
-    ) {
-      context
-        .select(
-          USER.EMAIL,
-          USER.NAME,
-          WORKFLOW_USER_ACCESS.PRIVILEGE
-        )
-        .from(WORKFLOW_USER_ACCESS)
-        .join(USER)
-        .on(USER.UID.eq(WORKFLOW_USER_ACCESS.UID))
-        .where(WORKFLOW_USER_ACCESS.WID.eq(wid).and(WORKFLOW_USER_ACCESS.UID.notEqual(user.getUid)))
-        .fetch()
-        .map(access => { access.into(classOf[AccessEntry2]) })
-        .toList
-    } else {
-      throw new ForbiddenException("You are not the owner of the workflow.")
-    }
+      .from(WORKFLOW_USER_ACCESS)
+      .join(USER)
+      .on(USER.UID.eq(WORKFLOW_USER_ACCESS.UID))
+      .where(
+        WORKFLOW_USER_ACCESS.WID
+          .eq(wid)
+          .and(WORKFLOW_USER_ACCESS.UID.notEqual(sessionUser.getUser.getUid))
+      )
+      .fetch()
+  }
+
+  /**
+    * This method shares a workflow to a user with a specific access type
+    *
+    * @param wid       the given workflow
+    * @param email     the email which the access is given to
+    * @param privilege the type of Access given to the target user
+    * @return rejection if user not permitted to share the workflow or Success Message
+    */
+  @PUT
+  @Path("/grant/{wid}/{email}/{privilege}")
+  def grantAccess(
+      @PathParam("wid") wid: UInteger,
+      @PathParam("email") email: String,
+      @PathParam("privilege") privilege: String
+  ): Unit = {
+    workflowUserAccessDao.merge(
+      new WorkflowUserAccess(
+        userDao.fetchOneByEmail(email).getUid,
+        wid,
+        WorkflowUserAccessPrivilege.valueOf(privilege)
+      )
+    )
   }
 
   /**
     * This method identifies the user access level of the given workflow
     *
-    * @param wid      the given workflow
+    * @param wid   the given workflow
     * @param email the email of the use whose access is about to be removed
     * @return message indicating a success message
     */
@@ -147,64 +157,15 @@ class WorkflowAccessResource() {
   @Path("/revoke/{wid}/{email}")
   def revokeAccess(
       @PathParam("wid") wid: UInteger,
-      @PathParam("email") email: String,
-      @Auth sessionUser: SessionUser
+      @PathParam("email") email: String
   ): Unit = {
-    if (
-      !workflowOfUserDao.existsById(
-        context
-          .newRecord(WORKFLOW_OF_USER.UID, WORKFLOW_OF_USER.WID)
-          .values(sessionUser.getUser.getUid, wid)
+    context
+      .delete(WORKFLOW_USER_ACCESS)
+      .where(
+        WORKFLOW_USER_ACCESS.UID
+          .eq(userDao.fetchOneByEmail(email).getUid)
+          .and(WORKFLOW_USER_ACCESS.WID.eq(wid))
       )
-    ) {
-      throw new ForbiddenException("No sufficient access privilege.")
-    } else {
-      context
-        .delete(WORKFLOW_USER_ACCESS)
-        .where(
-          WORKFLOW_USER_ACCESS.UID
-            .eq(userDao.fetchOneByEmail(email).getUid)
-            .and(WORKFLOW_USER_ACCESS.WID.eq(wid))
-        )
-        .execute()
-    }
-  }
-
-  /**
-    * This method shares a workflow to a user with a specific access type
-    * @param wid         the given workflow
-    * @param email    the email which the access is given to
-    * @param privilege the type of Access given to the target user
-    * @return rejection if user not permitted to share the workflow or Success Message
-    */
-  @PUT
-  @Path("/grant/{wid}/{email}/{privilege}")
-  @RolesAllowed(Array("REGULAR", "ADMIN"))
-  def grantAccess(
-      @PathParam("wid") wid: UInteger,
-      @PathParam("email") email: String,
-      @PathParam("privilege") privilege: String,
-      @Auth sessionUser: SessionUser
-  ): Unit = {
-    val uid: UInteger =
-      try {
-        userDao.fetchOneByEmail(email).getUid
-      } catch {
-        case _: IndexOutOfBoundsException =>
-          throw new BadRequestException("Target user does not exist.")
-      }
-    if (
-      !workflowOfUserDao.existsById(
-        context
-          .newRecord(WORKFLOW_OF_USER.UID, WORKFLOW_OF_USER.WID)
-          .values(sessionUser.getUser.getUid, wid)
-      )
-    ) {
-      throw new ForbiddenException("No sufficient access privilege.")
-    } else {
-      workflowUserAccessDao.merge(
-        new WorkflowUserAccess(uid, wid, WorkflowUserAccessPrivilege.valueOf(privilege))
-      )
-    }
+      .execute()
   }
 }
